requirements/develop.txt
# SHA1:242fd86f61f8954bb27c1d3a44732cf71fba1934
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
alabaster==0.7.12
    # via sphinx
asgiref==3.5.2
    # via django
asttokens==2.1.0
    # via stack-data
attrs==22.1.0
    # via pytest
babel==2.11.0
    # via sphinx
backcall==0.2.0
    # via ipython
black==22.10.0
    # via -r develop.in
build==0.9.0
    # via pip-tools
certifi==2022.12.7
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-compile-multi
    #   pip-tools
coverage[toml]==6.5.0
    # via
    #   -r develop.in
    #   pytest-cov
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
distlib==0.3.6
    # via virtualenv
django==3.2.16
    # via
    #   -r base.in
    #   django-appconf
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-import-export
    #   django-linear-migrations
    #   django-multiselectfield
    #   django-select2
    #   django-version-checks
django-appconf==1.0.5
    # via django-select2
django-debug-toolbar==3.7.0
    # via -r develop.in
django-extensions==3.2.1
    # via -r develop.in
django-filter==22.1
    # via -r base.in
django-import-export==3.0.1
    # via -r base.in
django-linear-migrations==2.5.1
    # via -r develop.in
django-multiselectfield==0.1.12
    # via -r base.in
django-select2==8.0.0
    # via -r base.in
django-version-checks==1.7.0
    # via -r base.in
docutils==0.19
    # via sphinx
et-xmlfile==1.1.0
    # via openpyxl
execnet==1.9.0
    # via pytest-xdist
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via -r develop.in
faker==15.2.0
    # via factory-boy
filelock==3.8.0
    # via virtualenv
flake8==5.0.4
    # via -r develop.in
identify==2.5.8
    # via pre-commit
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
iniconfig==1.1.1
    # via pytest
ipython==8.6.0
    # via -r develop.in
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via sphinx
./wheels/lxml-4.9.2-cp311-cp311-win_amd64.whl ; platform_system == "Windows"
lxml==4.9.2 ; platform_system != "Windows"
    # via python-docx
markuppy==1.14
    # via tablib
markupsafe==2.1.1
    # via jinja2
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
mypy-extensions==0.4.3
    # via black
nodeenv==1.7.0
    # via pre-commit
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
packaging==21.3
    # via
    #   build
    #   django-version-checks
    #   pytest
    #   sphinx
parso==0.8.3
    # via jedi
pathspec==0.10.1
    # via black
pep517==0.13.0
    # via build
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.3.0
    # via -r base.in
pip-compile-multi==2.5.0
    # via -r develop.in
pip-tools==6.9.0
    # via pip-compile-multi
pkgutil-resolve-name==1.3.10
    # via pytest-restrict
platformdirs==2.5.3
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.20.0
    # via -r develop.in
prompt-toolkit==3.0.32
    # via ipython
psycopg2-binary==2.9.5
    # via -r base.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pygments==2.13.0
    # via
    #   ipython
    #   sphinx
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via
    #   -r develop.in
    #   pytest-cov
    #   pytest-django
    #   pytest-env
    #   pytest-randomly
    #   pytest-restrict
    #   pytest-xdist
pytest-cov==4.0.0
    # via -r develop.in
pytest-django==4.5.2
    # via -r develop.in
pytest-env==0.8.1
    # via -r develop.in
pytest-randomly==3.12.0
    # via -r develop.in
pytest-restrict==4.4.0
    # via -r develop.in
pytest-xdist==3.0.2
    # via -r develop.in
python-dateutil==2.8.2
    # via
    #   -r base.in
    #   faker
python-docx==0.8.11
    # via -r base.in
python-dotenv==0.21.0
    # via -r base.in
pytz==2022.6
    # via
    #   babel
    #   django
pyyaml==6.0
    # via
    #   pre-commit
    #   tablib
requests==2.28.1
    # via sphinx
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
sphinx==5.3.0
    # via -r develop.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.0
    # via ipython
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
toml==0.10.2
    # via pre-commit
toposort==1.7
    # via pip-compile-multi
traitlets==5.5.0
    # via
    #   ipython
    #   matplotlib-inline
urllib3==1.26.12
    # via requests
virtualenv==20.16.6
    # via pre-commit
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.38.2
    # via pip-tools
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
